HW6/awesome-o/requirements.txt
# Install from repo root: pip install -r requirements.txt
pydantic>=2.6
pydantic-ai>=0.0.22
google-genai>=1.56.0
python-dotenv>=1.0
tqdm>=4.66

# Lecture video pipeline (PDF slides → Gemini → TTS → FFmpeg video)
google-generativeai>=0.8.0
pdf2image>=1.17.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
edge-tts>=7.0.0
